requirements.txt
Django==2.2.2
django-neomodel==0.0.4
djangorestframework==3.7.7
neo4j==1.7.4
neo4j-driver==1.7.2
neobolt==1.7.13
neomodel==3.3.2
neotime==1.7.4
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
whitenoise==3.3.1
